Include monitor by type, funding, GBD data
Lot of changes in the code this time. Country names have been standardized to
what is in the AQLI data in code (instead of excel as before). 3 out of the 7
data sources are neing read in directly from google sheets, where they are maintained,
to enable use of most up to date data. Accordingly, a few files from the
`oppotunity-score/data/input` folder have been deleted as the sheets' url is
preserved in code.

In addition, this iteration of the opportunity score has been refined with the
following changes:
   - Quintiles of population and PM2.5 ranging from 0.2 to 1 at intervals of 0.2
   - Countries where LYL due to PM2.5 is among top 10 cauases of death have been
   indicated to have greater opportunity
   - Places with funding of less than $100,000 have been indicated as having
   greater opportunity
   - Number of monitors have been split up by govt and non-govt monitors and been
   included as such